pyproject.toml
[tool.poetry]
name = "storage"
version = "0.1.0"
description = ""
authors = ["SALT Astronomy Operations <salt-software@saao.ac.za>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.61.2"}
python-dotenv = "^0.15.0"
databases = {extras = ["mysql"], version = "^0.4.1"}

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
flake8 = "^3.8.4"
flake8-docstrings = "^1.5.0"
black = "^20.8b1"
isort = "^5.6.4"
mypy = "^0.790"
uvicorn = "<0.12.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
